split postprocessing
