+++ Makefile
@@ -1,12 +1,8 @@
 SECONDARY_OUTPUT = output/demand_by_birth.png output/demand_by_time.csv
-# NB: main compute and post-processing are not separated here; it's run_scenarios.py for both
-POSTPROCESSING_SCRIPT = scripts/run_scenarios.py
 MAIN_OUTPUT = output/results.csv
-MAIN_SCRIPT = scripts/run_scenarios.py
 PACKAGE_CODE = drugdemand/__init__.py drugdemand/nirsevimab.py
 SCENARIOS = input/scenarios.yaml
-DATA_INPUT = input/births.csv input/weights.csv
-PREPROCESSING_SCRIPT = scripts/preprocess.py
+INPUT = input/births.csv input/weights.csv
 RAW_DATA = data/Natality,\ 2016-2022\ expanded.txt data/weights.csv data/hhs_regions.yaml
 
 .PHONY: clean
@@ -14,24 +10,24 @@ RAW_DATA = data/Natality,\ 2016-2022\ expanded.txt data/weights.csv data/hhs_reg
 all: $(SECONDARY_OUTPUT)
 
 # POSTPROCESSING --------------------------------------------------------------
-$(SECONDARY_OUTPUT): $(MAIN_OUTPUT) $(POSTPROCESSING_SCRIPT)
-	python $(POSTPROCESSING_SCRIPT)
+$(SECONDARY_OUTPUT): scripts/postprocess.py $(MAIN_OUTPUT) $(POSTPROCESSING_SCRIPT)
+	python $<
 
 # MAIN COMPUTATION ------------------------------------------------------------
-$(MAIN_OUTPUT): $(DATA_INPUTS) $(SCENARIOS) $(MAIN_SCRIPT) $(PACKAGE_CODE)
-	python $(MAIN_SCRIPT)
+$(MAIN_OUTPUT): scripts/run_scenarios.py $(INPUT) $(SCENARIOS) $(PACKAGE_CODE)
+	python $<
 
 # PREPROCESSING ---------------------------------------------------------------
-$(INPUTS): $(PREPROCESSING_SCRIPT) $(RAW_DATA)
-	python $(PREPROCESSING_SCRIPT)
+$(INPUT): scripts/preprocess.py $(RAW_DATA)
+	python $<
 
-$(SCENARIOS): scripts/write_scenarios.py
-	python scripts/write_scenarios.py
+$(SCENARIOS): scripts/create_scenarios.py
+	python $<
 
 # Data history ----------------------------------------------------------------
 # not intended to be run by the user; kept here for data history tracking
 data/weights.csv: scripts/pct_heavier_by_age.R
-	Rscript scripts/pct_heavier_by_age.R
+	Rscript $<
 
 clean:
 	rm -f input/* output/*
